Update CompositeOperations 0.8.6
- Update .make to use order-only prerequisites. From "Don’t use directories as normal prerequisites.":

"
It’s virtually never correct to use a directory as a normal prerequisite for a target. GNU make treats directories as any other file and rebuilds the target if the modification time of the directory has changed. However, modification times of directories change whenever a file is added, removed, or renamed in that directory... this is almost never what you want. It is useful to list directories as order-only prerequisites, however.
"

http://make.mad-scientist.net/papers/rules-of-makefiles/